refactor(frontend): update shadcn-vue and some small changes to header
