+++ packages/frontend/components.json
@@ -9,12 +9,13 @@
 		"cssVariables": true,
 		"prefix": ""
 	},
+	"iconLibrary": "lucide",
 	"aliases": {
 		"components": "@/components",
-		"composables": "@/composables",
 		"utils": "@/lib/utils",
 		"ui": "@/components/ui",
-		"lib": "@/lib"
+		"lib": "@/lib",
+		"composables": "@/composables"
 	},
-	"iconLibrary": "lucide"
+	"registries": {}
 }
