+++ packages/frontend/package.json
@@ -16,7 +16,7 @@
 		"axios": "^1.9.0",
 		"class-variance-authority": "^0.7.1",
 		"clsx": "^2.1.1",
-		"lucide-vue-next": "^0.511.0",
+		"lucide-vue-next": "^0.553.0",
 		"pinia": "^3.0.1",
 		"reka-ui": "^2.3.0",
 		"tailwind-merge": "^3.3.0",
