+++ packages/frontend/src/components/UI/card-custom/index.ts
@@ -1,0 +1,1 @@
+export { default as CardCustom } from "./CardUICustom.vue";
