--- packages/frontend/src/components/UI/card/index.ts
@@ -1,1 +1,0 @@
-export { default as Card } from "./CardUI.vue";
