+++ packages/frontend/src/components/UI/dropdown-menu/DropdownMenu.vue
@@ -1,10 +1,6 @@
 <script setup lang="ts">
-import {
-	DropdownMenuRoot,
-	type DropdownMenuRootEmits,
-	type DropdownMenuRootProps,
-	useForwardPropsEmits,
-} from "reka-ui";
+import type { DropdownMenuRootEmits, DropdownMenuRootProps } from "reka-ui";
+import { DropdownMenuRoot, useForwardPropsEmits } from "reka-ui";
 
 const props = defineProps<DropdownMenuRootProps>();
 const emits = defineEmits<DropdownMenuRootEmits>();
