+++ packages/frontend/src/components/UI/dropdown-menu/DropdownMenuContent.vue
@@ -1,13 +1,8 @@
 <script setup lang="ts">
+import type { DropdownMenuContentEmits, DropdownMenuContentProps } from "reka-ui";
 import type { HTMLAttributes } from "vue";
 import { reactiveOmit } from "@vueuse/core";
-import {
-	DropdownMenuContent,
-	type DropdownMenuContentEmits,
-	type DropdownMenuContentProps,
-	DropdownMenuPortal,
-	useForwardPropsEmits,
-} from "reka-ui";
+import { DropdownMenuContent, DropdownMenuPortal, useForwardPropsEmits } from "reka-ui";
 import { cn } from "@/lib/utils";
 
 const props = withDefaults(
