+++ packages/frontend/src/components/UI/dropdown-menu/DropdownMenuGroup.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
-import { DropdownMenuGroup, type DropdownMenuGroupProps } from "reka-ui";
+import type { DropdownMenuGroupProps } from "reka-ui";
+import { DropdownMenuGroup } from "reka-ui";
 
 const props = defineProps<DropdownMenuGroupProps>();
 </script>
