+++ packages/frontend/src/components/UI/dropdown-menu/DropdownMenuLabel.vue
@@ -1,7 +1,8 @@
 <script setup lang="ts">
+import type { DropdownMenuLabelProps } from "reka-ui";
 import type { HTMLAttributes } from "vue";
 import { reactiveOmit } from "@vueuse/core";
-import { DropdownMenuLabel, type DropdownMenuLabelProps, useForwardProps } from "reka-ui";
+import { DropdownMenuLabel, useForwardProps } from "reka-ui";
 import { cn } from "@/lib/utils";
 
 const props = defineProps<
