+++ packages/frontend/src/components/UI/dropdown-menu/DropdownMenuRadioGroup.vue
@@ -1,10 +1,6 @@
 <script setup lang="ts">
-import {
-	DropdownMenuRadioGroup,
-	type DropdownMenuRadioGroupEmits,
-	type DropdownMenuRadioGroupProps,
-	useForwardPropsEmits,
-} from "reka-ui";
+import type { DropdownMenuRadioGroupEmits, DropdownMenuRadioGroupProps } from "reka-ui";
+import { DropdownMenuRadioGroup, useForwardPropsEmits } from "reka-ui";
 
 const props = defineProps<DropdownMenuRadioGroupProps>();
 const emits = defineEmits<DropdownMenuRadioGroupEmits>();
