+++ packages/frontend/src/components/UI/dropdown-menu/DropdownMenuSeparator.vue
@@ -1,7 +1,8 @@
 <script setup lang="ts">
+import type { DropdownMenuSeparatorProps } from "reka-ui";
 import type { HTMLAttributes } from "vue";
 import { reactiveOmit } from "@vueuse/core";
-import { DropdownMenuSeparator, type DropdownMenuSeparatorProps } from "reka-ui";
+import { DropdownMenuSeparator } from "reka-ui";
 import { cn } from "@/lib/utils";
 
 const props = defineProps<
