+++ packages/frontend/src/components/UI/dropdown-menu/DropdownMenuSub.vue
@@ -1,10 +1,6 @@
 <script setup lang="ts">
-import {
-	DropdownMenuSub,
-	type DropdownMenuSubEmits,
-	type DropdownMenuSubProps,
-	useForwardPropsEmits,
-} from "reka-ui";
+import type { DropdownMenuSubEmits, DropdownMenuSubProps } from "reka-ui";
+import { DropdownMenuSub, useForwardPropsEmits } from "reka-ui";
 
 const props = defineProps<DropdownMenuSubProps>();
 const emits = defineEmits<DropdownMenuSubEmits>();
