+++ packages/frontend/src/components/UI/dropdown-menu/DropdownMenuSubContent.vue
@@ -1,12 +1,8 @@
 <script setup lang="ts">
+import type { DropdownMenuSubContentEmits, DropdownMenuSubContentProps } from "reka-ui";
 import type { HTMLAttributes } from "vue";
 import { reactiveOmit } from "@vueuse/core";
-import {
-	DropdownMenuSubContent,
-	type DropdownMenuSubContentEmits,
-	type DropdownMenuSubContentProps,
-	useForwardPropsEmits,
-} from "reka-ui";
+import { DropdownMenuSubContent, useForwardPropsEmits } from "reka-ui";
 import { cn } from "@/lib/utils";
 
 const props = defineProps<DropdownMenuSubContentProps & { class?: HTMLAttributes["class"] }>();
