+++ packages/frontend/src/components/UI/dropdown-menu/DropdownMenuSubTrigger.vue
@@ -1,8 +1,9 @@
 <script setup lang="ts">
+import type { DropdownMenuSubTriggerProps } from "reka-ui";
 import type { HTMLAttributes } from "vue";
 import { reactiveOmit } from "@vueuse/core";
 import { ChevronRight } from "lucide-vue-next";
-import { DropdownMenuSubTrigger, type DropdownMenuSubTriggerProps, useForwardProps } from "reka-ui";
+import { DropdownMenuSubTrigger, useForwardProps } from "reka-ui";
 import { cn } from "@/lib/utils";
 
 const props = defineProps<
