+++ packages/frontend/src/components/UI/dropdown-menu/DropdownMenuTrigger.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
-import { DropdownMenuTrigger, type DropdownMenuTriggerProps, useForwardProps } from "reka-ui";
+import type { DropdownMenuTriggerProps } from "reka-ui";
+import { DropdownMenuTrigger, useForwardProps } from "reka-ui";
 
 const props = defineProps<DropdownMenuTriggerProps>();
