+++ packages/frontend/src/components/UI/tooltip/TooltipContent.vue
@@ -1,14 +1,8 @@
 <script setup lang="ts">
+import type { TooltipContentEmits, TooltipContentProps } from "reka-ui";
 import type { HTMLAttributes } from "vue";
 import { reactiveOmit } from "@vueuse/core";
-import {
-	TooltipArrow,
-	TooltipContent,
-	type TooltipContentEmits,
-	type TooltipContentProps,
-	TooltipPortal,
-	useForwardPropsEmits,
-} from "reka-ui";
+import { TooltipArrow, TooltipContent, TooltipPortal, useForwardPropsEmits } from "reka-ui";
 import { cn } from "@/lib/utils";
 
 defineOptions({
