+++ packages/frontend/src/components/UI/tooltip/TooltipProvider.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
-import { TooltipProvider, type TooltipProviderProps } from "reka-ui";
+import type { TooltipProviderProps } from "reka-ui";
+import { TooltipProvider } from "reka-ui";
 
 const props = withDefaults(defineProps<TooltipProviderProps>(), {
 	delayDuration: 0,
