+++ packages/frontend/src/components/UI/tooltip/TooltipTrigger.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
-import { TooltipTrigger, type TooltipTriggerProps } from "reka-ui";
+import type { TooltipTriggerProps } from "reka-ui";
+import { TooltipTrigger } from "reka-ui";
 
 const props = defineProps<TooltipTriggerProps>();
 </script>
