+++ packages/frontend/src/components/UI/tooltip/index.ts
@@ -1,4 +1,4 @@
-export { default as Tooltip } from "./TooltipUi.vue";
+export { default as Tooltip } from "./Tooltip.vue";
 export { default as TooltipContent } from "./TooltipContent.vue";
 export { default as TooltipProvider } from "./TooltipProvider.vue";
 export { default as TooltipTrigger } from "./TooltipTrigger.vue";
